Made TS stricter and refractored stuff
