+++ src/helpers/misc.ts
@@ -1,6 +1,6 @@
 import path from "node:path";
 import fs from "node:fs/promises"
-import { Command, serverDBData } from "../types";
+import type { Command, serverDBData } from "../types";
 import * as config from '../../config.json';
 import { REST, Routes } from "discord.js";
